+++ temba/flows/tests.py
@@ -5406,12 +5406,8 @@ class FlowRunTest(TembaTest):
         run = FlowRun.create(self.flow, self.contact)
 
         # give our run some webhook data
-        event = WebHookEvent.objects.create(
+        WebHookEvent.objects.create(
             org=self.org, run=run, channel=self.channel, created_by=self.admin, modified_by=self.admin
-        )
-
-        WebHookResult.objects.create(
-            event=event, status_code=200, url="", created_by=self.admin, modified_by=self.admin
         )
 
         # our run go bye bye
@@ -6429,7 +6425,7 @@ class FlowsTest(FlowFileTest):
         # make sure triggering without a url fails properly
         WebHookEvent.trigger_flow_webhook(FlowRun.objects.all().first(), None, "", msg)
         result = WebHookResult.objects.all().order_by("-id").first()
-        self.assertIn("No webhook_url specified, skipping send", result.message)
+        self.assertIn("No webhook_url specified, skipping send", result.response)
 
     def test_validate_flow_definition(self):
 
@@ -9599,7 +9595,7 @@ class FlowMigrationTest(FlowFileTest):
 
         # run the flow to make sure the nodes are connected correctly
         flow = self.get_flow("migrate_to_11_7")
-        with patch("requests.api.request") as mock_request:
+        with patch("requests.Session.send") as mock_request:
             mock_request.return_value = MockResponse(200, "success")
 
             run, = flow.start([], [self.contact])
@@ -9609,12 +9605,12 @@ class FlowMigrationTest(FlowFileTest):
 
         # check webhook calls made in correct order
         self.assertEqual(len(mock_request.mock_calls), 5)
-        self.assertEqual(mock_request.mock_calls[0][1], ("get", "http://example.com/hook1"))
-        self.assertEqual(mock_request.mock_calls[0][2]["headers"], {"Header1": "Value1", "User-agent": "RapidPro"})
-        self.assertEqual(mock_request.mock_calls[1][1], ("post", "http://example.com/hook2"))
-        self.assertEqual(mock_request.mock_calls[2][1], ("get", "http://example.com/hook3"))
-        self.assertEqual(mock_request.mock_calls[3][1], ("get", "http://example.com/hook4"))
-        self.assertEqual(mock_request.mock_calls[4][1], ("get", "http://example.com/hook5"))
+        self.assertEqual(mock_request.mock_calls[0][1][0].url, "http://example.com/hook1")
+        self.assertEqual(mock_request.mock_calls[0][1][0].headers, {"Header1": "Value1", "User-agent": "RapidPro"})
+        self.assertEqual(mock_request.mock_calls[1][1][0].url, "http://example.com/hook2")
+        self.assertEqual(mock_request.mock_calls[2][1][0].url, "http://example.com/hook3")
+        self.assertEqual(mock_request.mock_calls[3][1][0].url, "http://example.com/hook4")
+        self.assertEqual(mock_request.mock_calls[4][1][0].url, "http://example.com/hook5")
 
         # all migrated actions create a unique result value
         run.refresh_from_db()