+++ temba/orgs/tests.py
@@ -22,7 +22,7 @@ from django.urls import reverse
 from django.utils import timezone
 
 from temba.airtime.models import AirtimeTransfer
-from temba.api.models import APIToken, Resthook
+from temba.api.models import APIToken, Resthook, WebHookEvent, WebHookResult
 from temba.archives.models import Archive
 from temba.campaigns.models import Campaign, CampaignEvent
 from temba.channels.models import Channel
@@ -299,6 +299,18 @@ class OrgDeleteTest(TembaTest):
             # we should be starting with some mock s3 objects
             self.assertEqual(5, len(self.mock_s3.objects))
 
+            # add in some webhook results
+            WebHookEvent.objects.create(
+                org=org,
+                event=WebHookEvent.TYPE_RELAYER_ALARM,
+                data=dict(),
+                created_by=self.admin,
+                modified_by=self.admin,
+            )
+            WebHookResult.objects.create(
+                org=self.org, url="http://foo.bar", request="GET http://foo.bar", status_code=200, response="zap!"
+            )
+
             # release our primary org
             org.release(immediately=immediately)
 
@@ -318,6 +330,9 @@ class OrgDeleteTest(TembaTest):
                 # our channels and org are gone too
                 self.assertFalse(Channel.objects.filter(org=org).exists())
                 self.assertFalse(Org.objects.filter(id=org.id).exists())
+
+                # as are our webhook events
+                self.assertFalse(WebHookEvent.objects.filter(org=org).exists())
             else:
 
                 org.refresh_from_db()
@@ -4179,7 +4194,7 @@ class StripeCreditsTest(TembaTest):
             def __init__(self):
                 self.throw = False
 
-            def all(self):
+            def list(self):
                 return dict_to_struct("MockCardData", dict(data=[MockCard(), MockCard()]))
 
             def create(self, card):